pyproject.toml
[project]
name = "smart-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "gradio>=5.36.2",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "uv>=0.7.20",
    "uvicorn[standard]>=0.35.0",
]
